Create module for stations methods
