+++ app/poros/station_basic.rb
@@ -1,4 +1,8 @@
+require "./app/models/application.rb"
+
 class StationBasic
+  include Stationable
+  
   attr_reader :id,
               :api_id,
               :name,
@@ -26,38 +30,5 @@ class StationBasic
     @state          = station_data[:state]
     @zip_code       = station_data[:zip]
     @is_favorited   = station_favorited?(station_data[:id], user_id)
-  end
-
-  def station_favorited?(station_api_id, user_id)
-    return "User token not provided" if user_id.nil?
-    User.find(user_id) # Check for users that do not exist to raise exception
-    return false if (station = Station.find_by(api_id: station_api_id)).nil?
-    if station && user_station = UserStation.find_by(station_id: station.id, user_id: user_id)
-      user_station.favorited?
-    else
-      false
-    end
-  end
-  
-  def status_finder(code)
-    if code.present?
-      if code == 'E'
-        'Available'
-      elsif code == 'P'
-        'Coming Soon'
-      elsif code == 'T'
-        'Temporarily Closed'
-      end
-    else
-      'Status Unavailable'
-    end
-  end
-
-  def set_ev_network(ev_network)
-    if ev_network.nil?
-      "Non-Networked"
-    else
-      ev_network
-    end
   end
 end
